requirements.txt
beautifulsoup4==4.11.1
lxml==4.9.0
pandas==1.1.5
requests==2.27.1
psycopg2-binary==2.9.3
